requirements.txt
py-clob-client>=0.18.0
streamlit>=1.32.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
